Events (and the Event Bus) - Lesson 7
Events can be used by any presenter and are more efficient than joining
presenters(making one big presenter). This is because events allow for
interactions between presenters without coupling the presenters(The
header presenter did not know who created the UserNotHappyEvent and the
child presenter did not know the header was using the event).