Client Place Manager- Lesson 8
An error page was created. It was shown how to have pages require
permission to be displayed. The error page was altered so that when they
were trying to access a page that required permission the user was asked
for a log in. It did not actually allow them to log in.  It was shown
how to redirect the user  to the page they were trying to access