Actions- Lesson 9 First  Part
We learned how to create and use an action. We so far have failed to
successfully protect against XSRF do to some unknown mistake or glitch.
The cookie does not load for the client even though we declared.
Commented out code trying to prevent XSRF. Right now the action works
because we set isSecured to false in GetFirst.java.